+++ src/utils/Crosshair.ts
@@ -1,3 +1,7 @@
+//import
+
+import { CrosshairColor } from "@valapi/lib";
+
 //interface
 
 namespace Crosshair {
@@ -21,6 +25,8 @@ namespace Crosshair {
         "0b"?: number;
         "0t"?: number;
         "0l"?: number;
+        "0v"?: number;
+        "0g"?: number;
         "0o"?: number;
         "0a"?: number;
         "0m"?: number;
@@ -31,6 +37,8 @@ namespace Crosshair {
         "1b"?: number;
         "1t"?: number;
         "1l"?: number;
+        "1v"?: number;
+        "1g"?: number;
         "1o"?: number;
         "1a"?: number;
         "1m"?: number;
@@ -68,7 +76,11 @@ namespace Crosshair {
     export interface Lines {
         isEnable: boolean, // not in valorant settings
         Opacity: number;
-        Length: number;
+        Length: { // base on valorant settings
+            Value: number;
+            isChain: boolean;
+            SecondValue: number;
+        };
         Thickness: number;
         Offset: number;
         MovementError: Crosshair.LinesError;
@@ -78,6 +90,7 @@ namespace Crosshair {
     export interface Crosshair {
         Primary: {
             Crosshair: {
+                isHexCrosshairColor: boolean, // not in valorant settings
                 CrosshairColor: string,
                 OutLine: {
                     isEnable: boolean, // not in valorant settings
@@ -98,6 +111,7 @@ namespace Crosshair {
         AimDownSights: {
             CopyPrimaryCrosshair: boolean,
             Crosshair: {
+                isHexCrosshairColor: boolean, // not in valorant settings
                 CrosshairColor: string,
                 OutLine: {
                     isEnable: boolean, // not in valorant settings
@@ -126,6 +140,7 @@ namespace Crosshair {
         };
         SniperScope: {
             CenterDot: {
+                isHexColor: boolean, // not in valorant settings
                 Color: string,
                 isEnable: boolean, // not in valorant settings
                 Opacity: number,
@@ -140,6 +155,7 @@ namespace Crosshair {
 const _defaultCrosshair: Crosshair.Crosshair = {
     Primary: {
         Crosshair: {
+            isHexCrosshairColor: false,
             CrosshairColor: "0",
             OutLine: {
                 isEnable: true,
@@ -157,7 +173,11 @@ const _defaultCrosshair: Crosshair.Crosshair = {
         InnerLines: {
             isEnable: true,
             Opacity: 0.8,
-            Length: 6,
+            Length: {
+                Value: 6,
+                isChain: false,
+                SecondValue: 6,
+            },
             Thickness: 2,
             Offset: 3,
             MovementError: {
@@ -172,7 +192,11 @@ const _defaultCrosshair: Crosshair.Crosshair = {
         OuterLines: {
             isEnable: true,
             Opacity: 0.35,
-            Length: 2,
+            Length: {
+                Value: 2,
+                isChain: false,
+                SecondValue: 2,
+            },
             Thickness: 2,
             Offset: 10,
             MovementError: {
@@ -188,6 +212,7 @@ const _defaultCrosshair: Crosshair.Crosshair = {
     AimDownSights: {
         CopyPrimaryCrosshair: true,
         Crosshair: {
+            isHexCrosshairColor: false,
             CrosshairColor: "0",
             OutLine: {
                 isEnable: true,
@@ -204,7 +229,11 @@ const _defaultCrosshair: Crosshair.Crosshair = {
         InnerLines: {
             isEnable: true,
             Opacity: 0.8,
-            Length: 6,
+            Length: {
+                Value: 6,
+                isChain: false,
+                SecondValue: 6,
+            },
             Thickness: 2,
             Offset: 3,
             MovementError: {
@@ -219,7 +248,11 @@ const _defaultCrosshair: Crosshair.Crosshair = {
         OuterLines: {
             isEnable: true,
             Opacity: 0.35,
-            Length: 2,
+            Length: {
+                Value: 2,
+                isChain: false,
+                SecondValue: 2,
+            },
             Thickness: 2,
             Offset: 10,
             MovementError: {
@@ -244,6 +277,7 @@ const _defaultCrosshair: Crosshair.Crosshair = {
     },
     SniperScope: {
         CenterDot: {
+            isHexColor: false,
             Color: "7",
             isEnable: true,
             Opacity: 0.75,
@@ -316,7 +350,7 @@ class Crosshair {
                 // upper case
                 
                 if (String(code).endsWith("FF")) {
-                    if (myJSON.endsWith(`"c":8,"u":`) || myJSON.endsWith(`"c":8,"t":`)) {
+                    if (myJSON.endsWith(`"c":${CrosshairColor.to["Custom"]},"u":`) || myJSON.endsWith(`"c":${CrosshairColor.to["Custom"]},"t":`)) {
                         // hex color
         
                         myJSON += `"${code}"`;
@@ -385,7 +419,7 @@ class Crosshair {
             // Color
 
             if (myCode.P.c) { // color
-                if (myCode.P.c == 8 && myCode.P.u) { //custom
+                if (myCode.P.c == CrosshairColor.to["Custom"] && myCode.P.u) { //custom
                     this.Primary.Crosshair.CrosshairColor = myCode.P.u;
                 } else {
                     this.Primary.Crosshair.CrosshairColor = String(myCode.P.c);
@@ -448,7 +482,13 @@ class Crosshair {
                 }
 
                 if (myCode.P["0l"]) { // length
-                    this.Primary.InnerLines.Length = myCode.P["0l"];
+                    this.Primary.InnerLines.Length.Value = myCode.P["0l"];
+                }
+
+                if (myCode.P["0g"] === 1 && myCode.P["0v"]) { //0g:1 = enable
+                    this.Primary.InnerLines.Length.isChain = true;
+
+                    this.Primary.InnerLines.Length.SecondValue = myCode.P["0v"];
                 }
 
                 if (myCode.P["0o"]) { // offset
@@ -497,7 +537,13 @@ class Crosshair {
                 }
 
                 if (myCode.P["1l"]) { // length
-                    this.Primary.OuterLines.Length = myCode.P["1l"];
+                    this.Primary.OuterLines.Length.Value = myCode.P["1l"];
+                }
+
+                if (myCode.P["1g"] === 1 && myCode.P["1v"]) { //1g:1 = enable
+                    this.Primary.OuterLines.Length.isChain = true;
+
+                    this.Primary.OuterLines.Length.SecondValue = myCode.P["1v"];
                 }
 
                 if (myCode.P["1o"]) { // offset
@@ -544,7 +590,7 @@ class Crosshair {
             // Color
 
             if (myCode.A.c) { // color
-                if (myCode.A.c == 8 && myCode.A.u) { //custom
+                if (myCode.A.c == CrosshairColor.to["Custom"] && myCode.A.u) { //custom
                     this.AimDownSights.Crosshair.CrosshairColor = myCode.A.u;
                 } else {
                     this.AimDownSights.Crosshair.CrosshairColor = String(myCode.A.c);
@@ -599,7 +645,13 @@ class Crosshair {
                 }
 
                 if (myCode.A["0l"]) { // length
-                    this.AimDownSights.InnerLines.Length = myCode.A["0l"];
+                    this.AimDownSights.InnerLines.Length.Value = myCode.A["0l"];
+                }
+
+                if (myCode.A["0g"] === 1 && myCode.A["0v"]) { //0g:1 = enable
+                    this.AimDownSights.InnerLines.Length.isChain = true;
+
+                    this.AimDownSights.InnerLines.Length.SecondValue = myCode.A["0v"];
                 }
 
                 if (myCode.A["0o"]) { // offset
@@ -648,7 +700,13 @@ class Crosshair {
                 }
 
                 if (myCode.A["1l"]) { // length
-                    this.AimDownSights.OuterLines.Length = myCode.A["1l"];
+                    this.AimDownSights.OuterLines.Length.Value = myCode.A["1l"];
+                }
+
+                if (myCode.A["1g"] === 1 && myCode.A["1v"]) { //1g:1 = enable
+                    this.AimDownSights.OuterLines.Length.isChain = true;
+
+                    this.AimDownSights.OuterLines.Length.SecondValue = myCode.A["1v"];
                 }
 
                 if (myCode.A["1o"]) { // offset
@@ -690,7 +748,7 @@ class Crosshair {
                 this.SniperScope.CenterDot.isEnable = false;
             } else {
                 if (myCode.S["c"]) { // color
-                    if (myCode.S["c"] == 8 && myCode.S["t"]) { //custom
+                    if (myCode.S["c"] == CrosshairColor.to["Custom"] && myCode.S["t"]) { //custom
                         this.SniperScope.CenterDot.Color = myCode.S["t"];
                     } else {
                         this.SniperScope.CenterDot.Color = String(myCode.S["c"]);
@@ -748,7 +806,9 @@ class Crosshair {
 
         // Color
 
-        if (this.Primary.Crosshair.CrosshairColor !== _defaultCrosshair.Primary.Crosshair.CrosshairColor) {
+        if (this.Primary.Crosshair.isHexCrosshairColor !== _defaultCrosshair.Primary.Crosshair.isHexCrosshairColor) {
+            myCode["P"] += `c;${CrosshairColor.to["Custom"]};u;${Number(this.Primary.Crosshair.isHexCrosshairColor)}`;
+        } else if (this.Primary.Crosshair.CrosshairColor !== _defaultCrosshair.Primary.Crosshair.CrosshairColor) {
             myCode["P"] += `c;${this.Primary.Crosshair.CrosshairColor};`;
         }
 
@@ -806,8 +866,17 @@ class Crosshair {
                 myCode["P"] += `0t;${this.Primary.InnerLines.Thickness};`;
             }
 
-            if (this.Primary.InnerLines.Length !== _defaultCrosshair.Primary.InnerLines.Length) {
-                myCode["P"] += `0l;${this.Primary.InnerLines.Length};`;
+            if (this.Primary.InnerLines.Length.Value !== _defaultCrosshair.Primary.InnerLines.Length.Value) {
+                myCode["P"] += `0l;${this.Primary.InnerLines.Length.Value};`;
+            }
+
+            if (this.Primary.InnerLines.Length.isChain !== _defaultCrosshair.Primary.InnerLines.Length.isChain) {
+
+                if (this.Primary.InnerLines.Length.SecondValue !== _defaultCrosshair.Primary.InnerLines.Length.SecondValue) {
+                    myCode["P"] += `0v;${this.Primary.InnerLines.Length.SecondValue};`;
+                }
+
+                myCode["P"] += `0g;${Number(this.Primary.InnerLines.Length.isChain)};`;
             }
 
             if (this.Primary.InnerLines.Offset !== _defaultCrosshair.Primary.InnerLines.Offset) {
@@ -853,8 +922,17 @@ class Crosshair {
                 myCode["P"] += `1t;${this.Primary.OuterLines.Thickness};`;
             }
 
-            if (this.Primary.OuterLines.Length !== _defaultCrosshair.Primary.OuterLines.Length) {
-                myCode["P"] += `1l;${this.Primary.OuterLines.Length};`;
+            if (this.Primary.OuterLines.Length.Value !== _defaultCrosshair.Primary.OuterLines.Length.Value) {
+                myCode["P"] += `1l;${this.Primary.OuterLines.Length.Value};`;
+            }
+
+            if (this.Primary.OuterLines.Length.isChain !== _defaultCrosshair.Primary.OuterLines.Length.isChain) {
+
+                if (this.Primary.OuterLines.Length.SecondValue !== _defaultCrosshair.Primary.OuterLines.Length.SecondValue) {
+                    myCode["P"] += `1v;${this.Primary.OuterLines.Length.SecondValue};`;
+                }
+
+                myCode["P"] += `1g;${Number(this.Primary.OuterLines.Length.isChain)};`;
             }
 
             if (this.Primary.OuterLines.Offset !== _defaultCrosshair.Primary.OuterLines.Offset) {
@@ -895,7 +973,9 @@ class Crosshair {
 
             // Color
 
-            if (this.AimDownSights.Crosshair.CrosshairColor !== _defaultCrosshair.AimDownSights.Crosshair.CrosshairColor) {
+            if (this.AimDownSights.Crosshair.isHexCrosshairColor !== _defaultCrosshair.AimDownSights.Crosshair.isHexCrosshairColor) {
+                myCode["A"] += `c;${CrosshairColor.to["Custom"]};u;${Number(this.AimDownSights.Crosshair.isHexCrosshairColor)}`;
+            } else if (this.AimDownSights.Crosshair.CrosshairColor !== _defaultCrosshair.AimDownSights.Crosshair.CrosshairColor) {
                 myCode["A"] += `c;${this.AimDownSights.Crosshair.CrosshairColor};`;
             }
 
@@ -947,8 +1027,17 @@ class Crosshair {
                     myCode["A"] += `0t;${this.AimDownSights.InnerLines.Thickness};`;
                 }
 
-                if (this.AimDownSights.InnerLines.Length !== _defaultCrosshair.AimDownSights.InnerLines.Length) {
-                    myCode["A"] += `0l;${this.AimDownSights.InnerLines.Length};`;
+                if (this.AimDownSights.InnerLines.Length.Value !== _defaultCrosshair.AimDownSights.InnerLines.Length.Value) {
+                    myCode["A"] += `0l;${this.AimDownSights.InnerLines.Length.Value};`;
+                }
+
+                if (this.AimDownSights.InnerLines.Length.isChain !== _defaultCrosshair.AimDownSights.InnerLines.Length.isChain) {
+
+                    if (this.AimDownSights.InnerLines.Length.SecondValue !== _defaultCrosshair.AimDownSights.InnerLines.Length.SecondValue) {
+                        myCode["A"] += `0v;${this.AimDownSights.InnerLines.Length.SecondValue};`;
+                    }
+    
+                    myCode["A"] += `0g;${Number(this.AimDownSights.InnerLines.Length.isChain)};`;
                 }
 
                 if (this.AimDownSights.InnerLines.Offset !== _defaultCrosshair.AimDownSights.InnerLines.Offset) {
@@ -994,8 +1083,17 @@ class Crosshair {
                     myCode["A"] += `1t;${this.AimDownSights.OuterLines.Thickness};`;
                 }
 
-                if (this.AimDownSights.OuterLines.Length !== _defaultCrosshair.AimDownSights.OuterLines.Length) {
-                    myCode["A"] += `1l;${this.AimDownSights.OuterLines.Length};`;
+                if (this.AimDownSights.OuterLines.Length.Value !== _defaultCrosshair.AimDownSights.OuterLines.Length.Value) {
+                    myCode["A"] += `1l;${this.AimDownSights.OuterLines.Length.Value};`;
+                }
+
+                if (this.AimDownSights.OuterLines.Length.isChain !== _defaultCrosshair.AimDownSights.OuterLines.Length.isChain) {
+
+                    if (this.AimDownSights.OuterLines.Length.SecondValue !== _defaultCrosshair.AimDownSights.OuterLines.Length.SecondValue) {
+                        myCode["A"] += `1v;${this.AimDownSights.OuterLines.Length.SecondValue};`;
+                    }
+    
+                    myCode["A"] += `1g;${Number(this.AimDownSights.OuterLines.Length.isChain)};`;
                 }
 
                 if (this.AimDownSights.OuterLines.Offset !== _defaultCrosshair.AimDownSights.OuterLines.Offset) {
@@ -1037,7 +1135,9 @@ class Crosshair {
             myCode["S"] += `d;${Number(this.SniperScope.CenterDot.isEnable)};`;
         } else {
 
-            if (this.SniperScope.CenterDot.Color !== _defaultCrosshair.SniperScope.CenterDot.Color) {
+            if (this.SniperScope.CenterDot.isHexColor !== _defaultCrosshair.SniperScope.CenterDot.isHexColor) {
+                myCode["S"] += `c;${CrosshairColor.to["Custom"]};t;${Number(this.SniperScope.CenterDot.isHexColor)}`;
+            } else if (this.SniperScope.CenterDot.Color !== _defaultCrosshair.SniperScope.CenterDot.Color) {
                 myCode["S"] += `c;${this.SniperScope.CenterDot.Color};`;
             }
 
@@ -1082,18 +1182,6 @@ class Crosshair {
     //static
 
     public static readonly Default: Crosshair.Crosshair = _defaultCrosshair;
-
-    public static readonly Color = {
-        "0": 'White',
-        "1": 'Green',
-        "2": 'Yellow Green',
-        "3": 'Green Yellow',
-        "4": 'Yellow',
-        "5": 'Cyan',
-        "6": 'Pink',
-        "7": 'Red',
-        "8": 'Custom'
-    };
 
     /**
      *